+++ autopush/websocket.py
@@ -101,10 +101,6 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
     def deferToThread(self, func, *args, **kwargs):
         """deferToThread helper that tracks defers outstanding"""
         d = deferToThread(func, *args, **kwargs)
-
-        def trapCancel(fail):
-            fail.trap(CancelledError)
-
         self._callbacks.append(d)
 
         def f(result):
@@ -112,22 +108,24 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
                 self._callbacks.remove(d)
             return result
         d.addBoth(f)
-        d.addErrback(trapCancel)
         return d
 
     def deferToLater(self, when, func, *args, **kwargs):
         """deferToLater helper that tracks defers outstanding"""
-        d = Deferred()
+        def cancel(d):
+            d._cancelled = True
 
-        def trapCancel(fail):
-            fail.trap(CancelledError)
-
-        d.addErrback(trapCancel)
+        d = Deferred(canceller=cancel)
+        d._cancelled = False
         self._callbacks.append(d)
 
         def f():
             if d in self._callbacks:
                 self._callbacks.remove(d)
+
+            # Don't run if the deferred was cancelled already
+            if d._cancelled:
+                return
             try:
                 result = func(*args, **kwargs)
                 d.callback(result)
@@ -135,6 +133,9 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
                 d.errback(failure.Failure())
         reactor.callLater(when, f)
         return d
+
+    def trap_cancel(self, fail):
+        fail.trap(CancelledError)
 
     def force_retry(self, func, *args, **kwargs):
         """Forcefully retry a function in a thread until it doesn't error
@@ -339,6 +340,7 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
         uaid = getattr(self, "uaid", None)
         self._shutdown_ran = True
         self._should_stop = True
+        self._check_notifications = False
         if uaid:
             self.cleanUp()
 
@@ -540,12 +542,14 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
 
         # Are we paused? Try again later.
         if self.paused:
-            self.deferToLater(1, self.process_notifications)
+            d = self.deferToLater(1, self.process_notifications)
+            d.addErrback(self.trap_cancel)
             return
 
         # Webpush with any outstanding storage-based must all be cleared
         if self.use_webpush and any(self.updates_sent.values()):
-            self.deferToLater(1, self.process_notifications)
+            d = self.deferToLater(1, self.process_notifications)
+            d.addErrback(self.trap_cancel)
             return
 
         # Are we already running?
@@ -556,15 +560,15 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
         self._check_notifications = False
         self._more_notifications = True
 
-        # Prevent repeat calls
         if self.use_webpush:
             d = self.deferToThread(self.ap_settings.message.fetch_messages,
                                    self.uaid)
         else:
             d = self.deferToThread(
                 self.ap_settings.storage.fetch_notifications, self.uaid)
-        d.addErrback(self.error_notifications)
         d.addCallback(self.finish_notifications)
+        d.addErrback(self.trap_cancel)
+        d.addErrback(self.error_notifications)
         self._notification_fetch = d
 
     def error_notifications(self, fail):
@@ -580,6 +584,7 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
         # Are we paused, try again later
         if self.paused:
             self.deferToLater(1, self.process_notifications)
+            return
 
         # Process notifications differently based on webpush style or not
         if self.use_webpush:
@@ -644,9 +649,6 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
                              ttl=notif["ttl"])
             )
             self.sendJSON(msg)
-
-        # Trigger for another fetch (won't run till notifs are cleared)
-        self.process_notifications()
 
     def _send_ping(self):
         """Helper for ping sending that tracks when the ping was sent"""
@@ -838,6 +840,11 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
         if resume:
             # Resume consuming ack's
             self.transport.resumeProducing()
+
+        # When using webpush, we don't check again if we have outstanding
+        # notifications
+        if self.use_webpush and any(self.updates_sent.values()):
+                return
 
         # Should we check again?
         if self._check_notifications or self._more_notifications: